refactor: extract drug class from pharmacy in proper file
