+++ pharmacy.test.ts
@@ -1,0 +1,10 @@
+import { Drug } from "./drug";
+import { Pharmacy } from "./pharmacy";
+
+describe("Pharmacy", () => {
+  it("should decrease the benefit and expiresIn", () => {
+    expect(new Pharmacy([new Drug("test", 2, 3)]).updateBenefitValue()).toEqual(
+      [new Drug("test", 1, 2)]
+    );
+  });
+});
